+++ Utilities/ocvalidate/ValidateMisc.c
@@ -346,8 +346,6 @@ CheckMiscSecurity (
   OC_UEFI_CONFIG    *UserUefi;
   BOOLEAN           IsAuthRestartEnabled;
   BOOLEAN           HasVSMCKext;
-  CONST CHAR8       *BootProtect;
-  BOOLEAN           IsRequestBootVarRoutingEnabled;
   CONST CHAR8       *AsciiDmgLoading;
   UINT32            ExposeSensitiveData;
   CONST CHAR8       *AsciiVault;
@@ -370,24 +368,6 @@ CheckMiscSecurity (
   if (IsAuthRestartEnabled && !HasVSMCKext) {
     DEBUG ((DEBUG_WARN, "Misc->Security->AuthRestart is enabled, but VirtualSMC is not loaded at Kernel->Add!\n"));
     ++ErrorCount;
-  }
-
-  BootProtect                    = OC_BLOB_GET (&UserMisc->Security.BootProtect);
-  IsRequestBootVarRoutingEnabled = UserUefi->Quirks.RequestBootVarRouting;
-  if (AsciiStrCmp (BootProtect, "None") != 0
-    && AsciiStrCmp (BootProtect, "Bootstrap") != 0
-    && AsciiStrCmp (BootProtect, "BootstrapShort") != 0) {
-    DEBUG ((DEBUG_WARN, "Misc->Security->BootProtect is borked (Can only be None, Bootstrap, or BootstrapShort)!\n"));
-    ++ErrorCount;
-  } else if (AsciiStrCmp (BootProtect, "Bootstrap") == 0
-    || AsciiStrCmp (BootProtect, "BootstrapShort") == 0) {
-    if (!IsRequestBootVarRoutingEnabled) {
-      DEBUG ((DEBUG_WARN, "Misc->Security->BootProtect is set to %a which requires UEFI->Quirks->RequestBootVarRouting to be enabled!\n", BootProtect));
-      ++ErrorCount;
-    }
-    //
-    // NOTE: RequestBootVarRouting requires OpenRuntime.efi, which will be checked in UEFI checker.
-    //
   }
 
   AsciiDmgLoading = OC_BLOB_GET (&UserMisc->Security.DmgLoading);